server: classify warnet as Server and run Flask without gunicorn
